# workloads/demo-service/openapi.yaml
openapi: 3.0.3
info:
  title: StoneTusker Customer Notification API
  version: 1.1.0
  description: >-
    Reference application used to demonstrate a browser UI, secure delivery,
    GitOps, observability, failure detection, rollback, and Backstage API
    discovery. The service records fictional notifications in memory and never
    sends email, SMS, or webhooks.
servers:
  - url: http://demo-service.demo-service-development.svc.cluster.local
    description: Development cluster service
  - url: http://localhost:8081
    description: Developer workstation after Kubernetes port-forward
paths:
  /:
    get:
      summary: Open the browser-based notification demo
      operationId: getApplicationUi
      responses:
        '200':
          description: HTML application
          content:
            text/html:
              schema:
                type: string
  /api/v1/status:
    get:
      summary: Return service and release metadata
      operationId: getServiceMetadata
      responses:
        '200':
          description: Service metadata
          headers:
            X-Correlation-ID:
              schema:
                type: string
            X-Service-Version:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceMetadata'
  /healthz:
    get:
      summary: Liveness endpoint
      operationId: getHealth
      responses:
        '200':
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /readyz:
    get:
      summary: Readiness endpoint
      operationId: getReadiness
      responses:
        '200':
          description: Service can receive traffic
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '503':
          description: Controlled readiness failure or unavailable dependency
  /api/v1/notifications:
    get:
      summary: Return recent in-memory notification records
      operationId: listNotifications
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Most recent notifications, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationRecord'
    post:
      summary: Accept a fictional notification request
      operationId: createNotification
      parameters:
        - in: header
          name: X-Demo-Request
          required: false
          schema:
            type: string
        - in: header
          name: X-Correlation-ID
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
      responses:
        '202':
          description: Notification accepted for the demonstration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationRecord'
        '422':
          description: Invalid request
        '500':
          description: Controlled demo error mode
  /api/v1/notifications/{notificationId}:
    get:
      summary: Return one in-memory notification record
      operationId: getNotification
      parameters:
        - in: path
          name: notificationId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationRecord'
        '404':
          description: Notification not found
components:
  schemas:
    Status:
      type: object
      required: [status]
      properties:
        status:
          type: string
    ServiceMetadata:
      type: object
      required: [service, environment, version, status]
      properties:
        service:
          type: string
        environment:
          type: string
        version:
          type: string
        status:
          type: string
    NotificationRequest:
      type: object
      required: [channel, recipient, message]
      properties:
        channel:
          type: string
          enum: [email, sms, webhook]
        recipient:
          type: string
          minLength: 3
          maxLength: 200
        message:
          type: string
          minLength: 1
          maxLength: 2000
    NotificationRecord:
      allOf:
        - $ref: '#/components/schemas/NotificationRequest'
        - type: object
          required: [id, state, accepted_at, correlation_id]
          properties:
            id:
              type: string
              format: uuid
            state:
              type: string
              example: accepted
            accepted_at:
              type: string
              format: date-time
            correlation_id:
              type: string